+++ cli/commands/metadata_reload.go
@@ -9,7 +9,7 @@ import (
 )
 
 func newMetadataReloadCmd(ec *cli.ExecutionContext) *cobra.Command {
-	opts := &metadataReloadOptions{
+	opts := &MetadataReloadOptions{
 		EC: ec,
 	}
 
@@ -26,28 +26,31 @@ func newMetadataReloadCmd(ec *cli.ExecutionContext) *cobra.Command {
   hasura metadata export --endpoint "<endpoint>"`,
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			opts.EC.Spin("Reloading metadata...")
-			err := opts.run()
-			opts.EC.Spinner.Stop()
-			if err != nil {
-				return errors.Wrap(err, "failed to reload metadata")
-			}
-			opts.EC.Logger.Info("Metadata reloaded")
-			return nil
+			return opts.runWithInfo()
 		},
 	}
 
 	return metadataReloadCmd
 }
 
-type metadataReloadOptions struct {
+type MetadataReloadOptions struct {
 	EC *cli.ExecutionContext
 }
 
-func (o *metadataReloadOptions) run() error {
+func (o *MetadataReloadOptions) runWithInfo() error {
+	o.EC.Spin("Reloading metadata...")
+	err := o.run()
+	o.EC.Spinner.Stop()
+	if err != nil {
+		return errors.Wrap(err, "failed to reload metadata")
+	}
+	o.EC.Logger.Info("Metadata reloaded")
+	return nil
+}
 
+func (o *MetadataReloadOptions) run() error {
 	var err error
-	metadataHandler := projectmetadata.NewHandlerFromEC(ec)
+	metadataHandler := projectmetadata.NewHandlerFromEC(o.EC)
 	_, err = metadataHandler.ReloadMetadata()
 	if err != nil {
 		return errors.Wrap(err, "Cannot reload metadata")